Move stable player ID key helpers to lib to drop cross-analytic coupling.
Fleet and the shared per-player stream hook no longer import scores-only state for connection key encoding.
